% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/schema.R
\name{list_table_schemas}
\alias{list_table_schemas}
\title{List the available table schemas}
\usage{
list_table_schemas(dir = system.file("schemas", package = "wxpipe"))
}
\arguments{
\item{dir}{Directory containing \verb{<table>.yml} schema files. Defaults to the
schemas installed with the package.}
}
\value{
A sorted character vector of table names.
}
\description{
List the available table schemas
}
\examples{
list_table_schemas()
}
