% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wxpipe-package.R
\docType{package}
\name{wxpipe-package}
\alias{wxpipe}
\alias{wxpipe-package}
\title{wxpipe: Extract-and-Load Pipeline for Salzburg Weather Observations and Forecasts}
\description{
Ingests daily station observations from the GeoSphere Austria Data Hub and numerical weather forecasts from Open-Meteo, lands the unmodified API payloads in Google Cloud Storage and loads tidy, append-only tables into BigQuery. Designed to run as a scheduled Cloud Run Job; downstream modelling (dbt) and serving (Plumber) live elsewhere.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/arohland/weather-forecast-checker}
}

}
\author{
\strong{Maintainer}: Andreas Rohland \email{andreas.rohland@posteo.de}

Authors:
\itemize{
  \item Andreas Rohland \email{andreas.rohland@posteo.de}
}

}
\keyword{internal}
